fix(db): restore V53 to its applied bytes, move the fold correction to V54
Prod would not boot after #141:

  FlywayValidateException: Migration checksum mismatch for migration version 53
  -> Applied to database : -699100394
  -> Resolved locally    : 2143289479

V53 had already been applied to the EC2 database during dev verification --
the dev and prod profiles resolve the same POSTGRES_HOST -- so editing it in
#141 to add the loser-vs-loser collision handling changed its checksum and
Flyway refused to start. Editing an applied migration is the mistake; this
fixes it forward rather than repairing the schema history by hand.

V53 is restored byte-for-byte to the version that ran. Verified locally by
recomputing Flyway's line-based CRC32: the restored file hashes to
-699100394, exactly the value the database holds.

V54 carries the correction instead, and takes a simpler route than the one
#141 used. Rather than enumerating collision shapes before an UPDATE, it
INSERTs the survivor's rows with ON CONFLICT DO NOTHING and then DELETEs the
loser's. That absorbs every collision -- the survivor already holding the
row, and two losers both holding it -- with no case analysis, which is what
V53 got wrong by copying PersonRepository.repointTags (a one-source-to-one-
target helper) onto an N-to-one fold.

V54 is a no-op against the live database: V53 completed there, only 2-row
groups existed, so the fold table is empty and the index already exists
(CREATE UNIQUE INDEX IF NOT EXISTS). It matters for a restore from a
pre-V53 dump.

V54FoldMigrationIntegrationTest replaces the V53-targeted one: it migrates
to 53, drops the index, seeds a three-row group with two losers on the same
content_id, then migrates to 54 and asserts no tag was lost. Verified to
fail with "duplicate key value violates unique constraint
content_image_people_pkey" when V54's fold is swapped for V53's approach.
